+++ src/models/Client.java
@@ -7,7 +7,7 @@ public class Client extends Person {
         super();
     }
 
-    public Client(Long id, String firstName, String lastName, String email, Advisor advisor) {
+    public Client(Integer id, String firstName, String lastName, String email, Advisor advisor) {
         super(id, firstName, lastName, email);
         this.advisor = advisor;
     }
